+++ src/Stryker.Core/Stryker.Core.UnitTest/Mutants/MutantOrchestratorTests.cs
@@ -1,23 +1,12 @@
 using Microsoft.CodeAnalysis.CSharp;
 using Shouldly;
-using Stryker.Core.InjectedHelpers;
-using Stryker.Core.Mutants;
-using Stryker.Core.Mutators;
-using Stryker.Core.Options;
 using System.Linq;
 using Xunit;
 
 namespace Stryker.Core.UnitTest.Mutants
 {
-    public class MutantOrchestratorTests
+    public class MutantOrchestratorTests : MutantOrchestratorTestsBase
     {
-        private readonly CsharpMutantOrchestrator _target;
-
-        public MutantOrchestratorTests()
-        {
-            _target = new CsharpMutantOrchestrator(options: new StrykerOptions(mutationLevel: MutationLevel.Complete.ToString()));
-        }
-
         [Fact]
         public void IfStatementsShouldBeNested()
         {
@@ -242,6 +231,28 @@ private bool Out(int test, Func<int, bool>lambda )
         }
 
         [Fact]
+        public void ShouldInitializeOutParametersInLocalFunction()
+        {
+            string source = @"void TestMethod()
+        {
+            void SomeLocalFunction(out string x)
+            {
+                var test3 = 2 + 5;
+                x = $""test{1 + test3}"";
+            };
+        }";
+            string expected = @"void TestMethod()
+        {
+             void SomeLocalFunction(out string x)
+            {
+{x= default(string);}                var test3 = (StrykerNamespace.MutantControl.IsActive(0)?2 - 5:2 + 5);
+                x = (StrykerNamespace.MutantControl.IsActive(1)?$"""":$""test{(StrykerNamespace.MutantControl.IsActive(2)?1 - test3:1 + test3)}"");
+            };    }";
+
+            ShouldMutateSourceToExpected(source, expected);
+        }
+
+        [Fact]
         public void ShouldMutateConditionalExpressionProperly()
         {
             string source = @"void TestMethod()
@@ -258,6 +269,28 @@ private bool Out(int test, Func<int, bool>lambda )
                 return (StrykerNamespace.MutantControl.IsActive(0)?string.Empty?.Any(x => !string.IsEmpty(x)):(StrykerNamespace.MutantControl.IsActive(1)?""Stryker was here!"":string.Empty)?.All(x => (StrykerNamespace.MutantControl.IsActive(2)?string.IsEmpty(x):!string.IsEmpty(x))));
         };
     }";
+
+            ShouldMutateSourceToExpected(source, expected);
+        }
+
+        [Fact]
+        public void ShouldAddDefaultReturnToLocalFunction()
+        {
+            string source = @"void TestMethod()
+        {
+            string SomeLocalFunction()
+            {
+                while(true);
+            };
+        }";
+            string expected = @"void TestMethod()
+        {
+            string SomeLocalFunction()
+            {
+                while((StrykerNamespace.MutantControl.IsActive(1)?!(true):(StrykerNamespace.MutantControl.IsActive(0)?false:true)));
+                return default(string);
+        };}";
+
 
             ShouldMutateSourceToExpected(source, expected);
         }
@@ -734,278 +767,6 @@ namespace TestApp
         }
 
         [Fact]
-        public void ShouldAddReturnDefaultToMethods()
-        {
-            string source = @"bool TestMethod()
-{
-    ;
-}";
-            string expected = @"bool TestMethod()
-{
-    ;
-    return default(bool);
-}";
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Fact]
-        public void ShouldAddReturnDefaultToConversion()
-        {
-            string source = @"public static explicit operator string(TestClass value)
-{;
-}";
-            string expected = @"public static explicit operator string(TestClass value)
-{;
-return default(string);
-}";
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Fact]
-        public void ShouldNotMutateConstDeclaration()
-        {
-            var source = @"void Test(){
-const string text = ""a""+""b""+""c"";}";
-            var expected = @"void Test(){
-const string text = ""a""+""b""+""c"";}";
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Fact]
-        public void ShouldAddReturnDefaultToAsyncMethods()
-        {
-            string source = @"public async Task<bool> TestMethod()
-{
-    ;
-}";
-            string expected = @"public async Task<bool> TestMethod()
-{
-    ;
-    return default(bool);
-}";
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Fact]
-        public void ShouldNotAddReturnDefaultToEnumerationMethods()
-        {
-            string source = @"public static IEnumerable<object> Extracting<T>(this IEnumerable<T> enumerable, string propertyName)
-      {
-        foreach (var o in enumerable)
-        {
-            yield return value;
-        }
-      }
-public static IEnumerable<object> Extracting<T>(this IEnumerable<T> enumerable)
-      {
-        yield break;
-      }";
-            string expected = @"public static IEnumerable<object> Extracting<T>(this IEnumerable<T> enumerable, string propertyName)
-      {
-        foreach (var o in enumerable)
-        {
-            yield return value;
-        }
-      }
-public static IEnumerable<object> Extracting<T>(this IEnumerable<T> enumerable)
-      {
-        yield break;
-      }";
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Fact]
-        public void ShouldAddReturnDefaultToAsyncWithFullNamespaceMethods()
-        {
-            string source = @"public async System.Threading.Tasks.Task<bool> TestMethod()
-{
-    ;
-}";
-            string expected = @"public async System.Threading.Tasks.Task<bool> TestMethod()
-{
-    ;
-    return default(bool);
-}";
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Fact]
-        public void ShouldNotAddReturnDefaultToAsyncTaskMethods()
-        {
-            string source = @"public async Task TestMethod()
-{
-    ;
-}";
-            string expected = @"public async Task TestMethod()
-{
-    ;
-}";
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Fact]
-        public void ShouldNotAddReturnDefaultToMethodsWithReturnTypeVoid()
-        {
-            string source = @"void TestMethod()
-{
-    ;
-}";
-            string expected = @"void TestMethod()
-{
-    ;
-}";
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Fact]
-        public void ShouldSkipStringsInSwitchExpression()
-        {
-            string source = @"string TestMethod()
-{
-    return input switch
-    {
-        ""test"" => ""test""
-    };
-}";
-            string expected = @"string TestMethod()
-{
-    return input switch
-    {
-        ""test"" => (StrykerNamespace.MutantControl.IsActive(0)?"""":""test""
-)    };
-}";
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Theory]
-        [InlineData("=> Value = \"Hello, World!\";")]
-        [InlineData("{Value = \"Hello, World!\";}")]
-        public void ShouldMutateStaticConstructor(string source)
-        {
-            source = @"class Test {
-static string Value { get; }
-static TestClass() " + source + "}";
-
-            var expected = @"class Test {
-static string Value { get; }
-static TestClass() {using(new StrykerNamespace.MutantContext()){Value = (StrykerNamespace.MutantControl.IsActive(0)?"""":""Hello, World!"");}}}";
-
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Fact]
-        public void ShouldMutateStaticPropertiesInArrowForm()
-        {
-            var source = @"class Test {
-static string Value => """";
-static TestClass(){}}";
-
-            var expected = @"class Test {
-static string Value => (StrykerNamespace.MutantControl.IsActive(0)?""Stryker was here!"":"""");
-static TestClass(){using(new StrykerNamespace.MutantContext()){}}}";
-
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Fact]
-        public void ShouldMutatePropertiesInArrowFormEvenWithComplexConstruction()
-        {
-            var source = @"class Test {
-string Value {get => Out(out var x)? ""empty"": """";}
-static TestClass(){}}";
-
-            var expected = @"class Test {
-string Value {get {if(StrykerNamespace.MutantControl.IsActive(0)){return!(Out(out var x))? ""empty"": """";}else{return Out(out var x)? (StrykerNamespace.MutantControl.IsActive(1)?"""":""empty""): (StrykerNamespace.MutantControl.IsActive(2)?""Stryker was here!"":"""");}}}
-static TestClass(){using(new StrykerNamespace.MutantContext()){}}}";
-
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Fact]
-        public void ShouldMutateStaticProperties()
-        {
-            var source = @"class Test {
-static string Value
-{
-    get { return ""TestDescription"";}
-    set { value = ""TestDescription"";}
-}
-static TestClass(){}}";
-
-            var expected = @"class Test {
-static string Value
-{
-    get { return (StrykerNamespace.MutantControl.IsActive(0)?"""":""TestDescription"");}
-    set { value = (StrykerNamespace.MutantControl.IsActive(1)?"""":""TestDescription"");}
-}
-static TestClass(){using(new StrykerNamespace.MutantContext()){}}}";
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-        [Fact]
-        public void ShouldMarkStaticMutationStaticInPropertiesInitializer()
-        {
-            var source = @"class Test {
-static string Value {get;} = """";}";
-
-            var expected = @"class Test {
-static string Value {get;} = (StrykerNamespace.MutantControl.IsActive(0)?""Stryker was here!"":"""");}";
-            ShouldMutateSourceToExpected(source, expected);
-            _target.Mutants.Count.ShouldBe(1);
-            _target.Mutants.First().IsStaticValue.ShouldBeTrue();
-        }
-
-        [Fact]
-        public void ShouldMutatePropertiesAsArrowExpression()
-        {
-            var source = @"class Test {
-string Value => Generator(out var x) ? """" :""test"";
-}";
-
-            var expected = @"class Test {
-string Value {get{if(StrykerNamespace.MutantControl.IsActive(0)){return!(Generator(out var x) )? """" :""test"";}else{return Generator(out var x) ? (StrykerNamespace.MutantControl.IsActive(1)?""Stryker was here!"":"""" ):(StrykerNamespace.MutantControl.IsActive(2)?"""":""test"");}}}
-}";
-            ShouldMutateSourceToExpected(source, expected);
-        }
-
-
-        [Fact]
-        public void ShouldControlLocalDeclarationMutationAtTheBlockLevel()
-        {
-            var source = @"public static string FormatPrettyByte(Int64 value)
-		{
-			string[] SizeSuffixes = { ""bytes"", ""KB"", ""MB"", ""GB"", ""TB"", ""PB"", ""EB"", ""ZB"", ""YB"" };
-
-			int mag = (int)Math.Log(value, 1024);
-			decimal adjustedSize = (decimal)value / (1L << (mag * 10));
-
-			return $""{adjustedSize:n1} {SizeSuffixes[mag]}"";
-		}
-";
-
-            var expected = @"public static string FormatPrettyByte(Int64 value)
-{if(StrykerNamespace.MutantControl.IsActive(0))		{
-			string[] SizeSuffixes = {};
-
-			int mag = (int)Math.Log(value, 1024);
-			decimal adjustedSize = (decimal)value / (1L << (mag * 10));
-
-			return $""{adjustedSize:n1} {SizeSuffixes[mag]}"";
-		}
-else		{
-			string[] SizeSuffixes = { (StrykerNamespace.MutantControl.IsActive(1)?"""":""bytes""), (StrykerNamespace.MutantControl.IsActive(2)?"""":""KB""), (StrykerNamespace.MutantControl.IsActive(3)?"""":""MB""), (StrykerNamespace.MutantControl.IsActive(4)?"""":""GB""), (StrykerNamespace.MutantControl.IsActive(5)?"""":""TB""), (StrykerNamespace.MutantControl.IsActive(6)?"""":""PB""), (StrykerNamespace.MutantControl.IsActive(7)?"""":""EB""), (StrykerNamespace.MutantControl.IsActive(8)?"""":""ZB""), (StrykerNamespace.MutantControl.IsActive(9)?"""":""YB"" )};
-
-			int mag = (int)Math.Log(value, 1024);
-			decimal adjustedSize = (StrykerNamespace.MutantControl.IsActive(10)?(decimal)value * (1L << (mag * 10)):(decimal)value / ((StrykerNamespace.MutantControl.IsActive(11)?1L >> (mag * 10):1L << ((StrykerNamespace.MutantControl.IsActive(12)?mag / 10:mag * 10)))));
-
-			return (StrykerNamespace.MutantControl.IsActive(13)?$"""":$""{adjustedSize:n1} {SizeSuffixes[mag]}"");
-		}
-return default(string);}}";
-            ShouldMutateSourceToExpected(source, expected);
-
-        }
-
-        [Fact]
         // test for issue #1386
         public void ShouldNotLeakMutationsAcrossDefinitions()
         {
@@ -1018,32 +779,6 @@ string Value => Generator(out var x) ? """" :""test"";
 int GetId(string input) {if(StrykerNamespace.MutantControl.IsActive(0)){return!(int.TryParse(input, out var result) )? result : 0;}else{return int.TryParse(input, out var result) ? result : 0;}}
 string Value {get{if(StrykerNamespace.MutantControl.IsActive(1)){return!(Generator(out var x) )? """" :""test"";}else{return Generator(out var x) ? (StrykerNamespace.MutantControl.IsActive(2)?""Stryker was here!"":"""" ):(StrykerNamespace.MutantControl.IsActive(3)?"""":""test"");}}}}}}";
             ShouldMutateSourceToExpected(source, expected);
-        }
-
-        private void ShouldMutateSourceToExpected(string actual, string expected)
-        {
-            actual = @"using System;
-using System.Collections.Generic;
-            using System.Text;
-namespace StrykerNet.UnitTest.Mutants.TestResources
-    {
-        class TestClass
-        {" + actual + "}}";
-
-            expected = @"using System;
-using System.Collections.Generic;
-            using System.Text;
-namespace StrykerNet.UnitTest.Mutants.TestResources
-    {
-        class TestClass
-        {" + expected + "}}";
-            var actualNode = _target.Mutate(CSharpSyntaxTree.ParseText(actual).GetRoot());
-            actual = actualNode.ToFullString();
-            actual = actual.Replace(CodeInjection.HelperNamespace, "StrykerNamespace");
-            actualNode = CSharpSyntaxTree.ParseText(actual).GetRoot();
-            var expectedNode = CSharpSyntaxTree.ParseText(expected).GetRoot();
-            actualNode.ShouldBeSemantically(expectedNode);
-            actualNode.ShouldNotContainErrors();
         }
     }
 }