+++ src/Stryker.Core/Stryker.Core/MutantFilters/IgnoredMethodMutantFilter.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Stryker.Core.Mutants;
 using Stryker.Core.Options;
@@ -12,49 +13,38 @@ namespace Stryker.Core.MutantFilters
     /// Checks if the mutants are part of ignored method calls.
     /// </summary>
     /// <seealso cref="IMutantFilter" />
-    public class IgnoredMethodMutantFilter : IMutantFilter
+    public sealed class IgnoredMethodMutantFilter : IMutantFilter
     {
         public string DisplayName => "method filter";
+        private readonly SyntaxTriviaRemover _triviaRemover = new SyntaxTriviaRemover();
 
-        public IEnumerable<Mutant> FilterMutants(IEnumerable<Mutant> mutants, ReadOnlyFileLeaf file, IStrykerOptions options)
-        {
-            if (!options.IgnoredMethods.Any())
+        public IEnumerable<Mutant> FilterMutants(IEnumerable<Mutant> mutants, ReadOnlyFileLeaf file, IStrykerOptions options) =>
+            options.IgnoredMethods.Any() ?
+                    mutants.Where(m => !IsPartOfIgnoredMethodCall(m.Mutation.OriginalNode, options)) :
+                    mutants;
+
+        private bool IsPartOfIgnoredMethodCall(SyntaxNode syntaxNode, IStrykerOptions options) =>
+            syntaxNode switch
             {
-                return mutants;
-            }
+                // Check if the current node is an invocation. This will also ignore invokable properties like `Func<bool> MyProp { get;}`
+                InvocationExpressionSyntax invocation => MatchesAnIgnoredMethod(_triviaRemover.Visit(invocation.Expression).ToString(), options),
 
-            return mutants.Where(m => !IsPartOfIgnoredMethodCall(m.Mutation.OriginalNode, options));
-        }
-
-        private bool IsPartOfIgnoredMethodCall(SyntaxNode syntaxNode, IStrykerOptions options)
-        {
-            switch (syntaxNode)
-            {
-                // Check if the current node is an invocation and the expression is a member
-                // This will also ignore invokable properties like `Func<bool> MyProp { get;}`
-                case InvocationExpressionSyntax invocation when invocation.Expression is MemberAccessExpressionSyntax member:
-                    return options.IgnoredMethods.Any(r => r.IsMatch(member.Name.ToString()));
-                // Check when conditional access
-                case InvocationExpressionSyntax invocation when invocation.Expression is MemberBindingExpressionSyntax member:
-                    return options.IgnoredMethods.Any(r => r.IsMatch(member.Name.ToString()));
-                // Check when direct identifier
-                case InvocationExpressionSyntax invocation when invocation.Expression is IdentifierNameSyntax member:
-                    return options.IgnoredMethods.Any(r => r.IsMatch(member.ToString()));
                 // Check if the current node is an object creation syntax (constructor invocation).
-                case ObjectCreationExpressionSyntax creation:
-                {
-                    var methodName = creation.Type + ".ctor";
-                    return options.IgnoredMethods.Any(r => r.IsMatch(methodName));
-                }
-            }
+                ObjectCreationExpressionSyntax creation => MatchesAnIgnoredMethod(_triviaRemover.Visit(creation.Type) + ".ctor", options),
 
-            // Traverse the tree upwards
-            if (syntaxNode.Parent != null)
-            {
-                return IsPartOfIgnoredMethodCall(syntaxNode.Parent, options);
-            }
+                // Traverse the tree upwards.
+                SyntaxNode node when node.Parent != null => IsPartOfIgnoredMethodCall(syntaxNode.Parent, options),
+                _ => false,
+            };
 
-            return false;
+        private static bool MatchesAnIgnoredMethod(string expressionString, IStrykerOptions options) => options.IgnoredMethods.Any(r => r.IsMatch(expressionString));
+
+        /// <summary>
+        /// Removes comments, whitespace, and other junk from a syntax tree.
+        /// </summary>
+        private sealed class SyntaxTriviaRemover : CSharpSyntaxRewriter
+        {
+            public override SyntaxTrivia VisitTrivia(SyntaxTrivia trivia) => default;
         }
     }
 }