+++ src/Stryker.Core/Stryker.Core/Mutants/NodeOrchestrators/ArrayInitializerOrchestrator.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
@@ -6,17 +6,14 @@ namespace Stryker.Core.Mutants.NodeOrchestrators
 {
     internal class ArrayInitializerOrchestrator : ExpressionSpecificOrchestrator<InitializerExpressionSyntax>
     {
-        protected override bool CanHandle(InitializerExpressionSyntax t)
-        {
-            return (t.Kind() == SyntaxKind.ArrayInitializerExpression && t.Expressions.Count > 0);
-        }
+        protected override bool CanHandle(InitializerExpressionSyntax t) => t.Kind() == SyntaxKind.ArrayInitializerExpression && t.Expressions.Count > 0;
 
         // mutations must be controlled at the statement level as those are not really expressions.
         protected override MutationContext StoreMutations(InitializerExpressionSyntax node,
             IEnumerable<Mutant> mutations,
             MutationContext context)
         {
-            context.StatementLevelControlledMutations.AddRange(mutations);
+            context.AddStatementLevel(mutations);
             return context;
         }