+++ src/Stryker.Core/Stryker.Core/Mutants/NodeOrchestrators/StaticConstructorOrchestrator.cs
@@ -1,4 +1,4 @@
-﻿using System.Linq;
+using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -12,10 +12,7 @@ namespace Stryker.Core.Mutants.NodeOrchestrators
     {
         protected override bool NewContext => true;
 
-        protected override bool CanHandle(ConstructorDeclarationSyntax t)
-        {
-            return t.Modifiers.Any(x => x.Kind() == SyntaxKind.StaticKeyword);
-        }
+        protected override bool CanHandle(ConstructorDeclarationSyntax t) => t.Modifiers.Any(x => x.Kind() == SyntaxKind.StaticKeyword);
 
         /// <inheritdoc/>
         /// <remarks>Injects a static marker used for coverage information; this implies converting
@@ -29,6 +26,7 @@ namespace Stryker.Core.Mutants.NodeOrchestrators
             {
                 return mutated;
             }
+
             if (mutated.ExpressionBody != null)
             {
                 // we need a body to place the marker