+++ src/Stryker.Core/Stryker.Core/Options/StrykerOptions.cs
@@ -388,13 +388,16 @@ namespace Stryker.Core.Options
             return mappedDiffIgnoreFiles;
         }
 
-        private static IEnumerable<Regex> ValidateIgnoredMethods(IEnumerable<string> methodPatterns)
-        {
-            foreach (var methodPattern in methodPatterns.Where(x => !string.IsNullOrEmpty(x)))
-            {
-                yield return new Regex("^" + Regex.Escape(methodPattern).Replace("\\*", ".*") + "$", RegexOptions.IgnoreCase);
-            }
-        }
+        /// <summary>
+        /// This method returns regexes that are equivalent to the provided method patterns.
+        /// In a method pattern the wildcard ('*') matches zero-or-more characters that are not periods ('.').
+        /// The regexes check from either the beginning of a method name or immediately after a period ('.'), whether the rest of the method name matches the method pattern.
+        /// </summary>
+        private static IEnumerable<Regex> ValidateIgnoredMethods(IEnumerable<string> methodPatterns) =>
+            methodPatterns
+                .Where(x => !string.IsNullOrEmpty(x))
+                .Select(methodPattern => new Regex("^(?:[^.]*\\.)*" + Regex.Escape(methodPattern).Replace("\\*", "[^.]*") + "$", RegexOptions.IgnoreCase))
+                .ToList();
 
         private OptimizationFlags ValidateMode(string mode)
         {
@@ -574,12 +577,25 @@ namespace Stryker.Core.Options
                     "The thresholds must be between 0 and 100");
             }
 
-            // ThresholdLow and ThresholdHigh can be the same value
-            if (thresholdBreak >= thresholdLow || thresholdLow > thresholdHigh)
+            if (thresholdLow > thresholdHigh)
             {
                 throw new StrykerInputException(
                     ErrorMessage,
-                    "The values of your thresholds are incorrect. Change `--threshold-break` to the lowest value and `--threshold-high` to the highest.");
+                    $"Threshold high must be higher than or equal to threshold low. Current high: {thresholdHigh}, low: {thresholdLow}.");
+            }
+
+            if (thresholdBreak > thresholdHigh)
+            {
+                throw new StrykerInputException(
+                    ErrorMessage,
+                    $"Threshold high must be higher than or equal to threshold break. Current high: {thresholdHigh}, break: {thresholdBreak}.");
+            }
+
+            if (thresholdBreak > thresholdLow)
+            {
+                throw new StrykerInputException(
+                    ErrorMessage,
+                    $"Threshold low must be higher than or equal to threshold break. Current low: {thresholdLow}, break: {thresholdBreak}.");
             }
 
             return new Threshold(thresholdHigh, thresholdLow, thresholdBreak);